회원가입 로직 수정

+++ src/main/java/com/vacation/platform/stayfinder/certify/entity/TermsUserAgreementId.java
@@ -10,8 +10,8 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TermsUserAgreementId implements Serializable {
-    private Integer id;
-    private Integer userId;
-    private Integer termsId;
-    private Integer version;
+    private Long id;
+    private Long userId;
+    private Long termsId;
+    private Long version;
 }
